+++ src/components/CommentList/CommentList.js
@@ -1,0 +1,28 @@
+import React, { useState } from 'react';
+import '../CommentList/CommentList.scss';
+
+const CommentList = props => {
+  const { text } = props;
+  const commentList = text.map((data, index) => (
+    <li key={index.toString()}>
+      <span className="screen_out">댓글 작성자: </span>
+      <a href="#none" className="link_user">
+        sumi_zz
+      </a>
+      <span className="screen_out">댓글 내용: </span>
+      {data}
+      <button className="btn_like" type="button">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          aria-label="좋아요"
+          viewBox="0 0 24 24"
+          className="img_g"
+        >
+          <path d="M16.792 3.904A4.989 4.989 0 0 1 21.5 9.122c0 3.072-2.652 4.959-5.197 7.222-2.512 2.243-3.865 3.469-4.303 3.752-.477-.309-2.143-1.823-4.303-3.752C5.141 14.072 2.5 12.167 2.5 9.122a4.989 4.989 0 0 1 4.708-5.218 4.21 4.21 0 0 1 3.675 1.941c.84 1.175.98 1.763 1.12 1.763s.278-.588 1.11-1.766a4.17 4.17 0 0 1 3.679-1.938m0-2a6.04 6.04 0 0 0-4.797 2.127 6.052 6.052 0 0 0-4.787-2.127A6.985 6.985 0 0 0 .5 9.122c0 3.61 2.55 5.827 5.015 7.97.283.246.569.494.853.747l1.027.918a44.998 44.998 0 0 0 3.518 3.018 2 2 0 0 0 2.174 0 45.263 45.263 0 0 0 3.626-3.115l.922-.824c.293-.26.59-.519.885-.774 2.334-2.025 4.98-4.32 4.98-7.94a6.985 6.985 0 0 0-6.708-7.218Z" />
+        </svg>
+      </button>
+    </li>
+  ));
+  return <ul>{commentList}</ul>;
+};
+export default CommentList;
